Chore: move hex color validator to color component
