+++ esphome/config_validation.py
@@ -915,16 +915,6 @@ def color_temperature(value):
         raise Invalid("Color temperature cannot be negative")
     return val
 
-def hex_color(value):
-    if not value.startswith("#"):
-        raise Invalid("Color must start with #")
-    value = value.lstrip('#')
-    if len(value) != 6:
-        raise Invalid("Color must have six digits")
-    try:
-        return (int(value[0:2], 16), int(value[2:3], 16), int(value[3:5], 16))
-    except ValueError:
-        raise Invalid("Color must be hexadecimal")
 
 def validate_bytes(value):
     value = string(value)
